requirements.txt
asgiref==3.3.4
cached-property==1.5.2
Django==3.2.4
django-url-filter==0.3.15
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
enum-compat==0.0.3
PyJWT==1.7.1
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
